Add Feijocards III
From http://www.vekn.net/forum/event-reports-and-twd/75329-twd-feijocards-iii

I ran the deck through Sutekh to normalise the card names, so
the format is quite different from the posted deck, but more consistent
with the other TWD entries.

I've assumed the "Pentex" entry in the original post is "Pentex(tm)
Subversion", as being the most reasonable interpretation.